% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr_functions_leg.R
\name{irr}
\alias{irr}
\title{IRR}
\usage{
irr(cf, rejectlm1 = TRUE, frequ = 1)
}
\arguments{
\item{cf}{are evenyly spaced cash flows delivered as a vector}

\item{rejectlm1}{rejects values less than -100% if TRUE (imaginary values automatically rejected)}

\item{frequ}{is the periodicity of the cash flow (1 for annual, 12 for monthly, etc)}
}
\description{
A polyroot based IRR function (included for curiosity, don't use, not reliable with real data)
}
\examples{
irr(c(-1,1.1))
}
\keyword{IRR}
