% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr_functions_leg.R
\name{irr.z365}
\alias{irr.z365}
\title{IRR daily data}
\usage{
irr.z365(cf.z)
}
\arguments{
\item{cf.z}{is the daily cash flow zoo object}
}
\description{
calculates an IRR using the irr function with sparse daily data
}
\examples{
irr.z365(zoo(c(-1,1.1),as.Date("2018-1-1")+c(0,180)))
}
\keyword{IRR}
