% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions_leg.R
\name{natozero}
\alias{natozero}
\title{na to zero}
\usage{
natozero(x)
}
\arguments{
\item{x}{is a vector}
}
\description{
convenience function converts na's to zero in a vector
}
\examples{
natozero(c(1,NA))
}
\keyword{Convenience}
